requirements.txt
curl_cffi>=0.7.0
python-dotenv>=1.0.0
mcp[cli]>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
